Use decimal instead of double and add unit tests

+++ BackendCodingChallenge/BackendCodingChallenge.UnitTest/CalculatorsTests/CoordinateDistanceCalculatorTests.cs
@@ -11,7 +11,7 @@ namespace BackendCodingChallenge.UnitTest.CalculatorsTests
         public void ComputeDistance_SameCoordinates_ReturnZero()
         {
             var coordinateDistanceCalculator = new CoordinateDistanceCalculator();
-            var distance = coordinateDistanceCalculator.ComputeDistance(33.45, -21.45, 33.45, -21.45);
+            var distance = coordinateDistanceCalculator.ComputeDistance((decimal) 33.45, (decimal) -21.45, (decimal) 33.45, (decimal) -21.45);
             
             Assert.AreEqual(0, distance);
         }
@@ -22,7 +22,7 @@ namespace BackendCodingChallenge.UnitTest.CalculatorsTests
             //Montreal(Lat:45.50884, Long:-73.58781), Ottawa(Lat:45.41117, Long:-75.69812)
             //Distance between both cities is 165km
             var coordinateDistanceCalculator = new CoordinateDistanceCalculator();
-            var distance = coordinateDistanceCalculator.ComputeDistance(45.50884, -73.58781, 45.41117, -75.69812);
+            var distance = coordinateDistanceCalculator.ComputeDistance((decimal) 45.50884, (decimal) -73.58781, (decimal) 45.41117, (decimal) -75.69812);
             
             Assert.AreEqual(165, Math.Round(distance));
         }
