+++ BackendCodingChallenge/BackendCodingChallenge/Calculators/Scores/IScoresCalculator.cs
@@ -6,12 +6,12 @@ namespace BackendCodingChallenge.Calculators.Scores
     public interface IScoresCalculator
     {
         /// <summary>
-        /// Returns the score between 0 and 0.5 using the BBF algorithm
-        /// Ref: https://www.cs.ubc.ca/~lowe/papers/ijcv04.pdf
+        /// Returns the score between 0 and 1 using a custom algorithm
         /// </summary>
-        /// <param name="distance"></param>
+        /// <param name="citiesModel"></param>
+        /// <param name="parametersModel"></param>
         /// <returns></returns>
-        public List<KeyValuePair<int, double>> GetCitiesScores(CitiesModel citiesModel, SuggestionsParametersModel parametersModel);
+        public List<KeyValuePair<int, decimal>> GetCitiesScores(CitiesModel citiesModel, SuggestionsParametersModel parametersModel);
 
 
     }
